+++ DellOpenManage/Classes/AlertPolicy.psm1
@@ -1,0 +1,12 @@
+
+Class AlertPolicy {
+    [Int]$Id
+    [String]$Name
+    [String]$Description
+    [Boolean]$Enabled
+    [Boolean]$DefaultPolicy
+    [Boolean]$State
+    [Boolean]$Visible
+    [String]$Owner
+    [PSCustomObject]$PolicyData
+}