+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.6.0]() - 2023-04-26
+### Added
+- New-OMEAlertPolicy
+- Get-OMEAlertPolicy
+- Enable-OMEAlertPolicy
+- Disable-OMEAlertPolicy
+
+### Fixed
+- Issue with Get-OMEAlert
+
 ## [3.5.2]() - 2023-04-26
 ### Fixed
 - Fixed bug with -FilterBy Id in Get-OMECatalog, Get-OMEConfigurationBaseline, Get-OMEDiscovery, Get-OMEFirmwareBaseline